Update bundled version of libsodium sources to version 1.0.17 + commit db6f43d2
This version exposes new low-level ed25519 API, such as non-clamped versions of
curve point multiplication and modular arithmetic operations.

This update also includes commit db6f43d2 to expose modular multiplication,
which is the only modular operation that didn't make it to 1.0.17. This commit
is applied with patch -t -p1, to only include ed25519 modular multiplication,
and drop ristretto255 which is not part of 1.0.17.