Tidy up mypy config
Now that all files are annotated, apply common settings globally so that
the overrides only apply to `nacl.bindings`.

In doing so, account for an `Any`-expression in `scrypt.py`. I must not
have correctly configured mypy for `nacl.pwhash` in #718: I think it
should have been `nacl.pwhash.*` instead of `nacl.pwhash`.