CI: build with Visual Studio 2022
libsodium provides v143 builds (built with Visual Studio 2022)
so we can use them and switch to building with VS2022 instead
of VS2019.

This will be helpful for future arm64 builds since libsodium only
provides v143 arm64 binaries.

While at it, make sure that "VsDevCmd.bat" is also called in the
test job and not just in the wheel build job, to make sure that
the same compiler setup is used in both places.