Test with ARM64 Windows and also build wheels (#864)
* CI: build with Visual Studio 2022

libsodium provides v143 builds (built with Visual Studio 2022)
so we can use them and switch to building with VS2022 instead
of VS2019.

This will be helpful for future arm64 builds since libsodium only
provides v143 arm64 binaries.

While at it, make sure that "VsDevCmd.bat" is also called in the
test job and not just in the wheel build job, to make sure that
the same compiler setup is used in both places.

* CI: build and test on Windows ARM64

GitHub now provides native arm64 windows runners using the
"windows-11-arm" image, so we can easily add it to the test matrix.

Python ARM64 builds are only available since 3.11, so only test with
Python 3.11 to 3.13.

The arm64 libsodium binaries have been added with c5253adde92b4a46faeb4

* CI: build wheels for Windows ARM64

Use the new native "windows-11-arm" Windows arm64 image for building.

Since arm64 Python is only available as of Python 3.11, use that for
building instead of 3.9, which is used for the other arches.